+++ lib/sidekiq_background_job.rb
@@ -23,7 +23,7 @@ module SidekiqBackgroundJob
       throw "Worker not implemented"
     end
 
-    def perform(job_id, *args)      
+    def perform(job_id, versions_metadata, *args)
       @progress_total = 1
       @progress = 0
       @job = BackgroundJob.find(job_id)
@@ -31,6 +31,9 @@ module SidekiqBackgroundJob
       return cancel_worker! if cancelling?
 
       ::PaperTrail.request.whodunnit = @job.user_id
+      ::PaperTrail.request.controller_info = versions_metadata.symbolize_keys.merge(
+        background_job_id: job_id
+      )
 
       # Execute original implementation of the background job worker.
       perform_job(*args)
@@ -40,6 +43,7 @@ module SidekiqBackgroundJob
       @job.fail!(error.message)
     ensure
       ::PaperTrail.request.whodunnit = nil
+      ::PaperTrail.request.controller_info = {}
     end
 
     def set_progress_total!(total)
@@ -68,20 +72,23 @@ module SidekiqBackgroundJob
     def cancel_worker!
       raise SidekiqBackgroundJob::Cancellation
     end
-    
+
     def cancelling?
       @job.reload
       @job.cancelling?
     end
 
     class << self
-      alias_method :perform_async_without_job, :perform_async 
+      alias_method :perform_async_without_job, :perform_async
       def perform_async(*args)
         options = args.extract_options!
 
-        background_job = BackgroundJob.create(options)
+        background_job = BackgroundJob.create(
+          { user_id: PaperTrail.request.whodunnit }.merge(options)
+        )
 
-        perform_async_without_job(background_job.id, *args)
+        versions_metadata = PaperTrail.request.controller_info || {}
+        perform_async_without_job(background_job.id, versions_metadata, *args)
 
         background_job
       end